Create README.md
ZingoMart is a full-stack eCommerce web app built with Django and PostgreSQL, offering a complete online shopping experience—from product browsing to secure checkout using Stripe. It includes user authentication, cart and order management, and an admin panel for backend control. Designed with clean UI and real-world flow, it's ideal for learning full-stack development or as a base for future eCommerce projects.